moving absence functions to absence file
